Refactor avg_weight [re]calculating
Each time we viewed the weights page all records for that user
would have their avg_weight recalculated.  This was starting to
impact performance, and is just generally messy.

The new structure only updates the record list on update and create.

When creating a record, just that record average is calculated

When updating a record, that record's average as well as any after
it are recalculated

TODO:  When creating or destroying a new record in the middle of a
record list all the averages will need to be recalculated above that
record in the list.  There has to be a more DRY way of doing this
on all the records when modifying a Weight object.  before_save looks
promising, but an initial attempt indicated it to be an infinite
recursion minefield.